do not create the file if already exists

+++ cliApp/fileCreator.js
@@ -30,9 +30,41 @@ class fileCreator {
 
   makeFile(file, filePath) {
     if (!fs.existsSync(this.directory)) {
-      fs.mkdirSync(this.directory, { recursive: true });
+      this.mkDirByPathSync(this.directory);
+    }
+    if (fs.existsSync(filePath)) {
+      return console.log(`ERROR: file: ${filePath} already exists.`);
     }
     return writeFile(filePath, file);
+  }
+
+  mkDirByPathSync(targetDir, { isRelativeToScript = false } = {}) {
+    const sep = path.sep;
+    const initDir = path.isAbsolute(targetDir) ? sep : '';
+    const baseDir = isRelativeToScript ? __dirname : '.';
+
+    return targetDir.split(sep).reduce((parentDir, childDir) => {
+      const curDir = path.resolve(baseDir, parentDir, childDir);
+      try {
+        fs.mkdirSync(curDir);
+      } catch (err) {
+        if (err.code === 'EEXIST') { // curDir already exists!
+          return curDir;
+        }
+
+      // To avoid `EISDIR` error on Mac and `EACCES`-->`ENOENT` and `EPERM` on Windows.
+        if (err.code === 'ENOENT') { // Throw the original parentDir error on curDir `ENOENT` failure.
+          throw new Error(`EACCES: permission denied, mkdir '${parentDir}'`);
+        }
+
+        const caughtErr = ['EACCES', 'EPERM', 'EISDIR'].indexOf(err.code) > -1;
+        if (!caughtErr || caughtErr && curDir === path.resolve(targetDir)) { //eslint-disable-line
+          throw err; // Throw if it's just the last created dir.
+        }
+      }
+
+      return curDir;
+    }, initDir);
   }
 
   compileString(string, values) {
